fix: rename hasCookie to index file
