+++ src/app/lib/utilities/hasCookie/index.ts
